requirements.txt
python-magic
tqdm
pandas
python-dotenv
pymongo
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client